requirements.txt
streamlit>=1.28.0
openpyxl>=3.1.2
pandas>=2.0.0
networkx>=3.1
streamlit-agraph>=0.0.45
openai>=1.0.0
google-generativeai>=0.3.0
pytest>=7.4.0
hypothesis>=6.88.0
ruff>=0.1.0
pre-commit>=3.5.0
playwright>=1.40.0
Pillow>=10.0.0
